# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SmartLibrary API",
        "description": "Dokumentasi API SmartLibrary",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/books": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Menampilkan semua daftar buku",
                "operationId": "a9d13283dfab262e97af3c914d51b7a8",
                "responses": {
                    "200": {
                        "description": "Sukses mengambil data buku"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Book"
                ],
                "summary": "Menambahkan data buku baru",
                "operationId": "47047e95a700b3271b697e57d0965191",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "author",
                                    "publish_year"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Laskar Pelangi"
                                    },
                                    "author": {
                                        "type": "string",
                                        "example": "Andrea Hirata"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Novel tentang perjuangan anak-anak Belitong."
                                    },
                                    "publish_year": {
                                        "type": "integer",
                                        "example": 2005
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Buku berhasil ditambahkan"
                    },
                    "400": {
                        "description": "Validasi gagal"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/books/{id}": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Melihat detail satu buku berdasarkan ID",
                "operationId": "2fdd668eb661591758aff4c56ca9b66a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID dari buku yang dicari",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data buku ditemukan"
                    },
                    "404": {
                        "description": "Buku tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Book"
                ],
                "summary": "Memperbarui data buku berdasarkan ID",
                "operationId": "f5ca658c501a803d68655af0365e3802",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID dari buku yang akan diubah",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Laskar Pelangi Edisi Revisi"
                                    },
                                    "author": {
                                        "type": "string",
                                        "example": "Andrea Hirata"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Deskripsi diperbarui."
                                    },
                                    "publish_year": {
                                        "type": "integer",
                                        "example": 2005
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Buku berhasil diperbarui"
                    },
                    "404": {
                        "description": "Buku tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Book"
                ],
                "summary": "Menghapus data buku berdasarkan ID",
                "operationId": "1a9c26667c32b8ed5e5770d7a1cc9d5c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID dari buku yang akan dihapus",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Buku berhasil dihapus"
                    },
                    "404": {
                        "description": "Buku tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/categories": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Menampilkan semua daftar kategori buku",
                "operationId": "3f5817a34833d0a1f4af4548dd3aeaba",
                "responses": {
                    "200": {
                        "description": "Sukses mengambil data kategori"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Menambahkan kategori baru",
                "operationId": "71fcad552bb0eaba9fb191fd8d8dcab0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Novel"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Kategori berhasil dibuat"
                    },
                    "400": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Melihat detail satu kategori",
                "operationId": "c68e76d323c008827a9e47398b1583de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kategori ditemukan"
                    },
                    "404": {
                        "description": "Kategori tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/books/{id}/category/{categoryId}": {
            "put": {
                "tags": [
                    "Category"
                ],
                "summary": "Menghubungkan buku dengan kategori tertentu (Many-to-Many)",
                "operationId": "f4b3d5b0bd1bfc831ff49d9e42523aee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Buku",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "ID Kategori",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kategori berhasil ditambahkan ke buku"
                    },
                    "404": {
                        "description": "Buku atau Kategori tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/books/{id}/categories": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Melihat daftar kategori yang dimiliki suatu buku",
                "operationId": "935d7c539acac17b39fb1c67e80e04fa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sukses mengambil kategori dari buku"
                    },
                    "404": {
                        "description": "Buku tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token JWT kamu di sini (tanpa kata 'Bearer ')",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}